+++ NEWS.md
@@ -1,3 +1,10 @@
+# ggtukey 0.2.0
+
+* Added additional options to `where` parameter for placement of letters.
+* Allowed numeric arguments to `type` argument in `geom_tukey`.
+* Added citation file.
+* Changed `vjust` default to `-0.2`.
+
 # ggtukey 0.1.0
 
 * Added a `NEWS.md` file to track changes to the package.